update key stroke
